+++ src/Reader.php
@@ -22,6 +22,7 @@ final class Reader
     /**
      * @return Attempt<Document|Node|Element|Custom>
      */
+    #[\NoDiscard]
     public function __invoke(Content $content): Attempt
     {
         $content = $content->toString();
@@ -30,23 +31,19 @@ final class Reader
             return Attempt::error(new \RuntimeException('Empty content'));
         }
 
-        $xml = new \DOMDocument;
-        /** @psalm-suppress ImpureMethodCall */
-        $success = $xml->loadXML(
-            $content,
-            \LIBXML_ERR_ERROR | \LIBXML_NOWARNING | \LIBXML_NOERROR,
-        );
-
-        if (!$success) {
-            return Attempt::error(new \RuntimeException('Failed to load xml content'));
+        try {
+            $xml = \Dom\XMLDocument::createFromString(
+                $content,
+                \LIBXML_ERR_ERROR | \LIBXML_NOWARNING | \LIBXML_NOERROR,
+            );
+        } catch (\Throwable $e) {
+            return Attempt::error($e);
         }
-
-        /** @psalm-suppress ImpureMethodCall */
-        $xml->normalizeDocument();
 
         return ($this->translate)($xml);
     }
 
+    #[\NoDiscard]
     public static function of(): self
     {
         return new self();