+++ src/Visitor/ParentNode.php
@@ -24,6 +24,7 @@ final class ParentNode
     /**
      * @return Maybe<Element|Custom>
      */
+    #[\NoDiscard]
     public function __invoke(Document|Node|Element|Custom $tree): Maybe
     {
         /** @var Maybe<Element|Custom> */
@@ -55,6 +56,7 @@ final class ParentNode
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(Node|Element|Custom $node): self
     {
         return new self($node);