+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - Requires PHP `8.4`
+- Requires `innmind/immutable:~6.0`
+- Requires `innmind/filesystem:~9.0`
 
 ## 8.0.0 - 2025-06-01